Added the frame filtering capabilities

+++ VirtualSense/DJ_VirtualMachine.1.0/darjeeling.1.1/src/lib/virtualsense/javax/virtualsense/network/Protocol.java
@@ -27,9 +27,9 @@ import javax.virtualsense.radio.Radio;
  *
  * @author Lattanzi
  */
-
+  
 public abstract class Protocol {
-    protected static short bestPath = -1;
+    protected static short bestPath =-1;
     private static boolean running = true;
     private static Packet actualPacket = null;
     private Semaphore packet = new Semaphore((short)0);
@@ -38,13 +38,9 @@ public abstract class Protocol {
     // invoke Radio.send(short, data) with bestPath as dest
     protected void send(Packet p){
         if(bestPath >= 0){
-        	//System.out.println("SEND UNICAST ");
-        	//System.out.println(bestPath);
-        	//Radio.broadcast(p.getData()); //Introduced to remove timeout for conference publication
-            Radio.send(bestPath, p.getData());     
+            Radio.send((short)bestPath, p.getData());     
         }else{
             Radio.broadcast(p.getData());
-            //System.out.println("SEND BROADCAST");
             
         }
     }
@@ -73,7 +69,6 @@ public abstract class Protocol {
         		short s_id = -1;
         		short r_id = -1;
         		Radio.init();
-        		//System.out.println(Thread.currentThread().getId());
         		while(running){
         			byte d[] = Radio.receive();
         			s_id = Radio.getSenderId();
